Update index.md
I thought this Resume was awesome! You have a lot of experience under your belt which is super helpful! All I thought of changing was moving the skills section above the service section because I think listing your programming skills is a little more prevalent than the mission for this resume.